+++ changelog.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### enhancements
+- clean weg items on load to remove duped items and apps/files that don't exist.
+
 ## [1.10.3]
 ### features
 - add beta channel